Move createUsers to regular helpers (no async/await)
