+++ test/api/helpers.js
@@ -1,7 +1,6 @@
 const request = require('supertest');
 const chai = require('chai');
 const app = require('../../app');
-const User = require('../../app/models/user');
 
 chai.should();
 
@@ -26,19 +25,3 @@ exports.testAdminResource = async (method, path) => {
     'You need to be an admin to access this resource.'
   );
 };
-
-const hash = '$2a$10$qxTI.cWwa2kwcjx4SI9KAuV4KxuhtlGOk33L999UQf1rux.4PBz7y'; // 'password'
-const testUser = (exports.testUser = {
-  username: 'testuser',
-  cardKey: '99TESTCARDKEY',
-  hash
-});
-
-const adminUser = (exports.adminUser = {
-  username: 'admin',
-  admin: true,
-  cardKey: '55TESTCARDKEY',
-  hash
-});
-
-exports.createUsers = async () => User.create([testUser, adminUser]);
